remove key module, collapse into signature
