+++ src/lib.rs
@@ -13,7 +13,6 @@ extern crate failure;
 
 mod encoding;
 mod errors;
-mod key;
 mod ops;
 mod point_ops;
 mod predicate;
